Lab1 Exploring Rust Data Structures
Objective: The objective of this lab is to
introduce students to the common collections
and data structures available in the Rust
programming language. By the end of the lab,
you should have a basic understanding of the
variety of collections and their uses.

See lab1_rust_data_structures/README.md for
more details.